move karma tests inline
