+++ src/index.ts
@@ -20,5 +20,6 @@
 export * from './language';
 export * from './util';
 export * from './component';
+export * from './form';
 export * from './formatter';
 export * from './type';
