+++ src/util/index.ts
@@ -4,4 +4,3 @@ export * from './promise-utils';
 export * from './storage-util';
 export * from './url-utils';
 export * from './data-utils';
-export * from './pagination-utils';
