--- tests/spec/component/pagination/pagination.spec.tsx
@@ -1,1 +1,0 @@
-/* TODO add unit tests */
